Add files via upload
This Java code is the solution for the below challenge

Tendable Tech Challenge
For this challenge, you'll need to upload your solution to GitHub (https://github.com/). If you don't
already have a GitHub account, please set one up before beginning.
Your task is to craft an automated test, preferably in C# or Java with Selenium or with any popular
language of your choice, to evaluate the following elements on the https://www.tendable.com
website:
1. Confirm accessibility of the top-level menus: Home, Our Story, Our Solution, and Why
Tendable.
2. Verify that the "Request a Demo" button is present and active on each of the
aforementioned top-level menu pages.
3. Navigate to the "Contact Us" section, choose "Marketing", and complete the form—
excluding the "Message" field. On submission, an error should arise. Ensure your script
confirms the error message's appearance. If the error is displayed, mark the test as PASS. If
absent, it's a FAIL.
4. In the event the test for point 3 fails (meaning no error message is displayed when the
"Message" field is empty), draft a bug report. This can be a Word document that you'll
upload to GitHub alongside your code.
5. Finally, incorporate a 'Readme.md' file in your submission detailing how to run your test
project, along with a brief description of: The strategy you employed for this challenge.